Move LowerTypes to its own file
